refactor(Stack): refactor stack contents
1. move all functions of Stack apply examples to a separate folder
2. build the separate folder `Example` as a library project
3. link the lib to its parent project and test project
