+++ src/CMakeLists.txt
@@ -7,7 +7,9 @@ include_directories(
 )
 
 add_subdirectory(DataStructure/Bitmap)
+add_subdirectory(DataStructure/Example)
 
 add_executable(dsa dsa.cpp)
 
 target_link_libraries(dsa bitmap)
+target_link_libraries(dsa example)
