+++ src/DataStructure/Example/CMakeLists.txt
@@ -1,0 +1,6 @@
+set(
+        SOURCE_FILES
+        StackExample.cpp
+)
+
+add_library(example ${SOURCE_FILES})
