+++ test/test_stack.cpp
@@ -1,4 +1,4 @@
-#include "Stack.hpp"
+#include "Example/StackExample.h"
 #include "catch2/catch.hpp"
